feat(encoder): indefinite data/text
Indefinite string support in decoder, but not in a standard way. Chunks
will be stored as separate CBOR objects. Alternative is to change String
structure to support data/text stored as separate non-continous chunks
and special String::read(...) fn to concatenate them afterwards in a
storage specified by user. But that would complicate everything.